LayoutTests/editing/selection/ios/do-not-allow-text-selection-in-user-select-none-expected.txt

This test verifies that long pressing an element with -webkit-user-select: none; does not begin a text selection gesture. To manually run the test, long press the square below and verify that no text is selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString() is ""
PASS successfullyParsed is true

TEST COMPLETE
Hello world
